requirements.txt
requests>=2.31.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
cryptography>=41.0.0
openai>=1.0.0
langchain-openai>=0.1.0
langchain>=0.1.0
numpy>=1.24.0
tqdm>=4.65.0
pydantic>=2.0.0
boto3>=1.34.0
botocore>=1.34.0
